release-kit|refactor: Extract `resolveCommandTags` helper
Extract the duplicated workspace-discovery, tag-resolution, and `--only` validation sequence from `publishCommand` and `githubReleaseCommand` into a shared `resolveCommandTags` helper. Also simplify the `=== true` comparison on `shouldInjectIntoReadme` and collapse `resolveReadmePath` to a single ternary expression.
