+++ packages/release-kit/src/publishCommand.ts
@@ -2,17 +2,16 @@
 /* eslint unicorn/no-process-exit: off */
 
 import { writeFileSync } from 'node:fs';
-import { basename, join } from 'node:path';
+import { join } from 'node:path';
 
 import { parseArgs, translateParseError } from '@williamthorsen/node-monorepo-core';
 
 import { createGithubReleases } from './createGithubRelease.ts';
 import { detectPackageManager } from './detectPackageManager.ts';
-import { discoverWorkspaces } from './discoverWorkspaces.ts';
 import { injectReleaseNotesIntoReadme, resolveReadmePath } from './injectReleaseNotesIntoReadme.ts';
 import { publishPackage } from './publish.ts';
+import { resolveCommandTags } from './resolveCommandTags.ts';
 import { resolveReleaseNotesConfig } from './resolveReleaseNotesConfig.ts';
-import { resolveReleaseTags } from './resolveReleaseTags.ts';
 
 const publishFlagSchema = {
   dryRun: { long: '--dry-run', type: 'boolean' as const },
@@ -37,43 +36,7 @@ export async function publishCommand(argv: string[]): Promise<void> {
   const { dryRun, noGitChecks, provenance } = parsed.flags;
   const only = parsed.flags.only?.split(',');
 
-  // Discover workspaces to determine single-package vs monorepo mode.
-  let discoveredPaths: string[] | undefined;
-  try {
-    discoveredPaths = await discoverWorkspaces();
-  } catch (error: unknown) {
-    console.error(`Error discovering workspaces: ${error instanceof Error ? error.message : String(error)}`);
-    process.exit(1);
-  }
-
-  if (only !== undefined && discoveredPaths === undefined) {
-    console.error('Error: --only is only supported for monorepo configurations');
-    process.exit(1);
-  }
-
-  // Build workspace map: dir (basename) -> workspace path.
-  const workspaceMap =
-    discoveredPaths === undefined ? undefined : new Map(discoveredPaths.map((p) => [basename(p), p]));
-
-  // Resolve tags from HEAD.
-  let resolvedTags = resolveReleaseTags(workspaceMap);
-
-  if (resolvedTags.length === 0) {
-    console.error('Error: No release tags found on HEAD. Create tags with `release-kit tag` first.');
-    process.exit(1);
-  }
-
-  // Validate --only against resolved tags.
-  if (only !== undefined) {
-    const availableNames = resolvedTags.map((t) => t.dir);
-    for (const name of only) {
-      if (!availableNames.includes(name)) {
-        console.error(`Error: Unknown package "${name}" in --only. Available: ${availableNames.join(', ')}`);
-        process.exit(1);
-      }
-    }
-    resolvedTags = resolvedTags.filter((t) => only.includes(t.dir));
-  }
+  const resolvedTags = await resolveCommandTags(only);
 
   if (resolvedTags.length === 0) {
     return;
@@ -82,7 +45,7 @@ export async function publishCommand(argv: string[]): Promise<void> {
   const packageManager = detectPackageManager();
   const { releaseNotes, changelogJsonOutputPath } = await resolveReleaseNotesConfig();
 
-  const shouldInject = releaseNotes.shouldInjectIntoReadme === true;
+  const shouldInject = releaseNotes.shouldInjectIntoReadme;
 
   // Print confirmation listing before publishing.
   console.info(dryRun ? '[dry-run] Would publish:' : 'Publishing:');
